ThermalGuardian/requirements.txt
anyio==4.6.2
astor==0.8.1
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==20.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
decorator==5.2.1
distro==1.4.0
entrypoints==0.3
exceptiongroup==1.3.0
filelock==3.16.1
fsspec==2025.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.14.0
httpx==0.28.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
isc==2.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linecache2==1.0.0
MarkupSafe==1.1.0
more-itertools==4.2.0
mpmath==1.3.0
netifaces==0.10.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
opt-einsum==3.3.0
paddle2onnx==1.3.1
paddlepaddle-gpu==2.6.2
pbr==5.4.5
pexpect==4.6.0
pillow==10.4.0
pip==20.0.2
ply==3.11
protobuf==3.20.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycurl==7.43.0.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
setuptools==45.2.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
sos==4.4
ssh-import-id==5.10
sympy==1.13.3
systemd-python==234
testresources==2.0.0
torch==2.4.0
traceback2==1.4.0
triton==3.0.0
Twisted==18.9.0
typing-extensions==4.13.2
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.25.8
wadllib==1.3.3
wheel==0.34.2
zipp==1.0.0
zope.interface==4.7.1
colorlog
easydict
numba >= 0.56
nuscenes-devkit
opencv-python <= 4.6.0.66
opencv-python-headless <= 4.6.0.66
pandas
paddledet
paddleseg
pyquaternion
rarfile
scikit-image
scikit-learn
visualdl
h5py
einops
albumentations
mmdeploy==0.14.0
mmcv==1.7.2
onnxruntime==1.19.2
